+++ CHANGELOG.md
@@ -6,6 +6,7 @@ All notable changes to this project will be documented in this file.
 ## Unreleased
 
 * Fix `DelegateProxy` call to `layoutIfNeeded` for an object without a window. #2076
+* Add `placeholder` property to `UITextField`. #2076
 
 ## [5.1.1](https://github.com/ReactiveX/RxSwift/releases/tag/5.1.1)